+++ src/main.py
@@ -1,0 +1,6 @@
+from GerenciadorTarefas import GerenciadorTarefas
+from InterfaceTarefa import InterfaceTarefa
+
+if __name__ == "__main__":
+    gerenciador = GerenciadorTarefas("assets/Tarefas/Tarefas.txt")
+    InterfaceTarefa(gerenciador)
